requirements.txt
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
celery==4.0.2
coreapi==2.1.1
coreschema==0.0.4
Django==1.10.5
django-celery-results==1.0.1
django-filter==1.0.1
django-logging-json==1.5.3
django-nose==1.4.4
django-rest-swagger==2.1.1
djangorestframework==3.5.3
elasticsearch==2.4.1
itypes==1.1.0
Jinja2==2.9.5
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==0.23
nose==1.3.7
nose2==0.6.5
openapi-codec==1.3.1
packaging==16.8
pika==0.10.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
urllib3==1.20
vine==1.1.3
